// components/extension-kit/src/sync/base-database.interface.ts
/**
 * Базовый интерфейс данных из базы данных для всех синхронизируемых сущностей
 */
export interface IBaseDatabaseData {
  /** Внутренний ID базы данных */
  _id: string;
  /** Номер блока последнего обновления */
  block_num?: number;
  /** Существует ли запись в блокчейне */
  present: boolean;
  /** Статус сущности */
  status?: string;
  /** Дата создания в базе данных */
  _created_at?: Date;
  /** Дата последнего обновления в базе данных */
  _updated_at?: Date;

}